almost all code updated

+++ docker-build-kernel.sh
@@ -1,21 +1,23 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
-# Find current directory
 SCRIPTPATH=$( cd $(dirname $0) ; pwd -P )
+# Source basic functions
+source "$SCRIPTPATH"/utils.sh
+
 INPUTDIR="$SCRIPTPATH""/origin"
 OUTPUTDIR="$SCRIPTPATH""/destination"
 
-echo "Input dir: $INPUTDIR"
-echo "Output dir: $OUTPUTDIR"
+print_details "Input dir:" "$INPUTDIR"
+print_details "Output dir:" "$OUTPUTDIR"
 
 if $(docker image inspect stockmind/gpd-pocket-ubuntu-respin:latest >/dev/null 2>&1); then
-	echo "Found Docker Hub image!"
+	print_info "Found Docker Hub image!"
 	IMAGENAME="stockmind/gpd-pocket-ubuntu-respin"
 elif $(docker image inspect gpd-pocket-ubuntu-respin:latest >/dev/null 2>&1); then
-	echo "Found local image!"
+	print_info "Found local image!"
 	IMAGENAME="gpd-pocket-ubuntu-respin"
 else
-	echo "Build docker image or download it from Docker Hub!"
+	print_error "Build docker image or download it from Docker Hub!"
 	exit 1
 fi
 
@@ -25,6 +27,12 @@ if [ "$1" != 'keepconfig' ]; then
 fi
 
 # Refresh container
-docker rm $(docker ps -aq --filter name=gpd-pocket-kernel-container)
+OLDCONTAINER=$(docker ps -aq --filter name=gpd-pocket-kernel-container)
+if [ $OLDCONTAINER != "" ]; then
+	docker rm $(docker ps -aq --filter name=gpd-pocket-kernel-container)
+fi
+
+IMAGENAME="gpd-pocket-ubuntu-respin-test"
+
 # Run command
 docker run -t -v "$INPUTDIR":/docker-input -v "$OUTPUTDIR":/docker-output --name gpd-pocket-kernel-container "$IMAGENAME" kernel
